✨ feat: new static system
