+++ app_test.go
@@ -668,253 +668,6 @@ func Test_App_Shutdown(t *testing.T) {
 	})
 }
 
-// go test -run Test_App_Static_Index_Default
-func Test_App_Static_Index_Default(t *testing.T) {
-	app := New()
-
-	app.Static("/prefix", "./.github/workflows")
-	app.Static("", "./.github/")
-	app.Static("test", "", Static{Index: "index.html"})
-
-	resp, err := app.Test(httptest.NewRequest(MethodGet, "/", nil))
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextHTMLCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-	body, err := io.ReadAll(resp.Body)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, true, strings.Contains(string(body), "Hello, World!"))
-
-	resp, err = app.Test(httptest.NewRequest(MethodGet, "/not-found", nil))
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 404, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextPlainCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-	body, err = io.ReadAll(resp.Body)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, "Cannot GET /not-found", string(body))
-}
-
-// go test -run Test_App_Static_Index
-func Test_App_Static_Direct(t *testing.T) {
-	app := New()
-
-	app.Static("/", "./.github")
-
-	resp, err := app.Test(httptest.NewRequest(MethodGet, "/index.html", nil))
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextHTMLCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-	body, err := io.ReadAll(resp.Body)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, true, strings.Contains(string(body), "Hello, World!"))
-
-	resp, err = app.Test(httptest.NewRequest(MethodGet, "/testdata/testRoutes.json", nil))
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMEApplicationJSON, resp.Header.Get("Content-Type"))
-	utils.AssertEqual(t, "", resp.Header.Get(HeaderCacheControl), "CacheControl Control")
-
-	body, err = io.ReadAll(resp.Body)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, true, strings.Contains(string(body), "testRoutes"))
-}
-
-// go test -run Test_App_Static_MaxAge
-func Test_App_Static_MaxAge(t *testing.T) {
-	app := New()
-
-	app.Static("/", "./.github", Static{MaxAge: 100})
-
-	resp, err := app.Test(httptest.NewRequest("GET", "/index.html", nil))
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, "text/html; charset=utf-8", resp.Header.Get(HeaderContentType))
-	utils.AssertEqual(t, "public, max-age=100", resp.Header.Get(HeaderCacheControl), "CacheControl Control")
-}
-
-// go test -run Test_App_Static_Download
-func Test_App_Static_Download(t *testing.T) {
-	app := New()
-	c := app.AcquireCtx(&fasthttp.RequestCtx{})
-	defer app.ReleaseCtx(c)
-
-	app.Static("/fiber.png", "./.github/testdata/fs/img/fiber.png", Static{Download: true})
-
-	resp, err := app.Test(httptest.NewRequest("GET", "/fiber.png", nil))
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, "image/png", resp.Header.Get(HeaderContentType))
-	utils.AssertEqual(t, `attachment`, resp.Header.Get(HeaderContentDisposition))
-}
-
-func Test_App_Static_Wildcard(t *testing.T) {
-	app := New()
-
-	app.Static("*", "./.github/index.html")
-
-	req := httptest.NewRequest(MethodGet, "/yesyes/john/doe", nil)
-	resp, err := app.Test(req)
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextHTMLCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-	body, err := io.ReadAll(resp.Body)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, true, strings.Contains(string(body), "Test file"))
-}
-
-func Test_App_Static_Prefix_Wildcard(t *testing.T) {
-	app := New()
-
-	app.Static("/test/*", "./.github/index.html")
-
-	req := httptest.NewRequest(MethodGet, "/test/john/doe", nil)
-	resp, err := app.Test(req)
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextHTMLCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-	app.Static("/my/nameisjohn*", "./.github/index.html")
-
-	resp, err = app.Test(httptest.NewRequest(MethodGet, "/my/nameisjohn/no/its/not", nil))
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextHTMLCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-	body, err := io.ReadAll(resp.Body)
-	utils.AssertEqual(t, nil, err)
-	utils.AssertEqual(t, true, strings.Contains(string(body), "Test file"))
-}
-
-func Test_App_Static_Prefix(t *testing.T) {
-	app := New()
-	app.Static("/john", "./.github")
-
-	req := httptest.NewRequest(MethodGet, "/john/index.html", nil)
-	resp, err := app.Test(req)
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextHTMLCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-	app.Static("/prefix", "./.github/testdata")
-
-	req = httptest.NewRequest(MethodGet, "/prefix/index.html", nil)
-	resp, err = app.Test(req)
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextHTMLCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-	app.Static("/single", "./.github/testdata/testRoutes.json")
-
-	req = httptest.NewRequest(MethodGet, "/single", nil)
-	resp, err = app.Test(req)
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMEApplicationJSON, resp.Header.Get(HeaderContentType))
-}
-
-func Test_App_Static_Trailing_Slash(t *testing.T) {
-	app := New()
-	app.Static("/john", "./.github")
-
-	req := httptest.NewRequest(MethodGet, "/john/", nil)
-	resp, err := app.Test(req)
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextHTMLCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-	app.Static("/john_without_index", "./.github/testdata/fs/css")
-
-	req = httptest.NewRequest(MethodGet, "/john_without_index/", nil)
-	resp, err = app.Test(req)
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 404, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextPlainCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-	app.Static("/john/", "./.github")
-
-	req = httptest.NewRequest(MethodGet, "/john/", nil)
-	resp, err = app.Test(req)
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextHTMLCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-	req = httptest.NewRequest(MethodGet, "/john", nil)
-	resp, err = app.Test(req)
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 200, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextHTMLCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-	app.Static("/john_without_index/", "./.github/testdata/fs/css")
-
-	req = httptest.NewRequest(MethodGet, "/john_without_index/", nil)
-	resp, err = app.Test(req)
-	utils.AssertEqual(t, nil, err, "app.Test(req)")
-	utils.AssertEqual(t, 404, resp.StatusCode, "Status code")
-	utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-	utils.AssertEqual(t, MIMETextPlainCharsetUTF8, resp.Header.Get(HeaderContentType))
-}
-
-func Test_App_Static_Next(t *testing.T) {
-	app := New()
-	app.Static("/", ".github", Static{
-		Next: func(c *Ctx) bool {
-			// If value of the header is any other from "skip"
-			// c.Next() will be invoked
-			return c.Get("X-Custom-Header") == "skip"
-		},
-	})
-	app.Get("/", func(c *Ctx) error {
-		return c.SendString("You've skipped app.Static")
-	})
-
-	t.Run("app.Static is skipped: invoking Get handler", func(t *testing.T) {
-		req := httptest.NewRequest("GET", "/", nil)
-		req.Header.Set("X-Custom-Header", "skip")
-		resp, err := app.Test(req)
-		utils.AssertEqual(t, nil, err)
-		utils.AssertEqual(t, 200, resp.StatusCode)
-		utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-		utils.AssertEqual(t, MIMETextPlainCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-		body, err := io.ReadAll(resp.Body)
-		utils.AssertEqual(t, nil, err)
-		utils.AssertEqual(t, true, strings.Contains(string(body), "You've skipped app.Static"))
-	})
-
-	t.Run("app.Static is not skipped: serving index.html", func(t *testing.T) {
-		req := httptest.NewRequest("GET", "/", nil)
-		req.Header.Set("X-Custom-Header", "don't skip")
-		resp, err := app.Test(req)
-		utils.AssertEqual(t, nil, err)
-		utils.AssertEqual(t, 200, resp.StatusCode)
-		utils.AssertEqual(t, false, resp.Header.Get(HeaderContentLength) == "")
-		utils.AssertEqual(t, MIMETextHTMLCharsetUTF8, resp.Header.Get(HeaderContentType))
-
-		body, err := io.ReadAll(resp.Body)
-		utils.AssertEqual(t, nil, err)
-		utils.AssertEqual(t, true, strings.Contains(string(body), "Hello, World!"))
-	})
-}
-
 // go test -run Test_App_Mixed_Routes_WithSameLen
 func Test_App_Mixed_Routes_WithSameLen(t *testing.T) {
 	app := New()
@@ -925,7 +678,7 @@ func Test_App_Mixed_Routes_WithSameLen(t *testing.T) {
 		return c.Next()
 	})
 	// routes with the same length
-	app.Static("/tesbar", "./.github")
+	app.Use("/tesbar", NewStatic("./.github"))
 	app.Get("/foobar", func(c *Ctx) error {
 		c.Type("html")
 		return c.Send([]byte("FOO_BAR"))
